backend/serverless/video_generation/requirements.txt
psycopg2-binary
numpy
google-cloud-storage
google-cloud-pubsub
google-cloud-secret-manager
google-cloud-run
